Switch to report table
Before this the report (and compare) view used the reports stored in the
AsyncResult which was retrieve (I believe) from celery (task_meta?!) via
the NmapTask table.
Now reports are copied into a "reports" table. The meta date (e.g. scan
task comment and so on are stored in a seperate table (nmap_report_meta)
which has a 1:1 relation to the reports table.